Add InMemoryDbContext for unit testing with in-memory DB
Introduced an abstract class `InMemoryDbContext` in the
`ClassLibrary.EFCore.Tests.DatabaseContext` namespace to simplify
the creation of in-memory database contexts for unit tests.

The `GetDbContext` method was added to return a `UnitTestDbContext`
configured with an in-memory database named "InMemory-Database-Test".
This change improves test isolation and eliminates the need for
an actual database during testing.

Imported `Microsoft.EntityFrameworkCore` to support the in-memory
database functionality.